requirements.txt
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
python-dotenv==0.20.0
